requirements.txt
flask>=2.0.0
scikit-learn>=1.0.0
nltk>=3.6.0
python-docx>=0.8.11
PyPDF2>=3.0.0
Werkzeug>=2.0.0 